third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/inline-scope-variables-expected.txt

CONSOLE MESSAGE: line 13: 3
Tests that inline scope variables are rendering correctly.

Set timer for test function.
9: length = 123
10: a = (3) [1, 2, 3]
11: 
12: b = 42
13: a = (3) [1, 2, 3]
